Extract Playlist from PlaylistPage
